requirements.txt
uvicorn
python-dotenv
fastapi
llama-index
llama-index-agent-openai
llama-index-llms-groq